+++ src/components/Form/Form.stories.tsx
@@ -988,7 +988,8 @@ const queryData: FormData<AllFieldsData> = {
   number: 1237,
   currency: 100,
   email: "email@email.com",
-  textarea: "Textarea",
+  textarea:
+    "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.",
   search: sharedOptions[0].value,
   searchText: "Search",
 };